requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
colorama>=0.4.6
anthropic>=0.32.0
requests